refactor(coordinator): 抽出 hotkey_supervisors / voice_agent_hotkeys 子模块
机械式行为保持拆分：从 coordinator.rs 搬出听写/QA/combo/翻译/action/window
热键的 supervisor/bridge/事件循环，以及 Less Computer 语音 Agent 热键簇。
保持 use super::* 重导出契约（pub(crate) use 子模块::*），Inner/ActionHotkeyKind
随之提升为 pub(crate)。
